Fix very large mistake
slit_pdf previously was un-normalised, and was then normalised within rv_from_pdf. This results in rv_from_pdf failing when given a normalised PDF, and slit_pdf would not behave as expected. This defines a new un-normalised function, intensity, which is then normalised in slit_pdf, and rv_from_pdf now expects a normalised PDF.